src/modules/audio/AudioManager.test.ts
import { AudioManager } from "../../../src/modules/audio/AudioManager";

jest.mock("speaker");

describe("AudioManager", () => {
  let audioManager: AudioManager;

  beforeEach(() => {
    audioManager = new AudioManager();
  });

  afterEach(() => {
    audioManager.cleanup();
  });

  describe("initialization", () => {
    it("should create an instance with default config", () => {
      expect(audioManager).toBeDefined();
    });
  });

  describe("playback", () => {
    it("should handle empty buffer gracefully", () => {
      audioManager.playVoice(Buffer.from([]));
      // TODO: Add assertions
    });

    it("should clean up resources on stop", () => {
      audioManager.playVoice(Buffer.from([]));
      audioManager.stopVoice();

      // TODO: Add assertions
    });
  });

  describe("cleanup", () => {
    it("should clean up all resources", () => {
      audioManager.playVoice(Buffer.from([]));
      audioManager.cleanup();

      // TODO: Add assertions
    });

    it("should handle multiple cleanup calls safely", () => {
      audioManager.cleanup();
      audioManager.cleanup();

      // TODO: Add assertions
    });
  });
});
